Strict enum mobs -> interface-based mobs
The idea here is to move away from the strictly enum-based approach
to a monster interface. The fallback feature makes it possible to
spawn any new members of EntityType without updating MACreature with
a new member.

As an example, assume a new monster type, monkey, is added to the
game. Bukkit will implement it as EntityType.MONKEY, but MACreature
won't be updated yet. The current approach still allows for 'monkey'
to be parsed as the correct EntityType.

Nothing is tested, and there's no telling if this will allow for any
kind of special mobs from the big mod packs or not (probably not,
since the entity IDs are internal to Bukkit, and don't necessarily
correspond to anything in Minecraft itself (speculations).